use cb.io for now
until ovy manages to fabricate our own Brain for the AI, we'll borrow cb.io's
