+++ src/commands/chatbot/package.json
@@ -1,6 +1,7 @@
 {
     "main": "chatbot.js",
     "dependencies": {
-        "markov-strings": "^1.3.2"
+        "markov-strings": "^1.3.2",
+        "better-cleverbot-io": "1.0.2"
     }
 }
